+++ src/api/CortexApi.ts
@@ -15,6 +15,7 @@
  */
 import {
   EntitySyncProgress,
+  ExpirationResponse,
   GroupByOption,
   Initiative,
   InitiativeActionItem,
@@ -106,4 +107,6 @@ export interface CortexApi {
   getUserPermissions(): Promise<UserPermissionsResponse>;
 
   getSyncJobs(): Promise<JobsResponse>;
+
+  getExpiration(): Promise<ExpirationResponse>;
 }